libstdc++: add a workaround for format_kind<optional<T>> [PR120644]
The specialization of format_kind for optional is causing a problem when
optional is imported and included. The comments on the PR strongly
suggest that this is a frontend bug; this commit just works around the
issue by specifying the type of format_kind<optional<T>> to be
`range_format`, rather than leaving the compiler deduce it via `auto`.

	PR c++/120644

libstdc++-v3/ChangeLog:

	* include/std/optional (format_kind): Do not use `auto`.