Fortran: fix checking of MOLD= in ALLOCATE statements [PR51961]
In ALLOCATE statements where the MOLD= argument is present and is not
scalar, and the allocate-object has an explicit-shape-spec, the standard
does not require the ranks to agree.  In that case we skip the rank check,
but emit a warning if -Wsurprising is given.

	PR fortran/51961

gcc/fortran/ChangeLog:

	* resolve.cc (conformable_arrays): Use modified rank check when
	MOLD= expression is given.

gcc/testsuite/ChangeLog:

	* gfortran.dg/allocate_with_mold_5.f90: New test.