Add GUESSED_GLOBAL0_AFDO
This patch adds GUESSED_GLOBAL0_AFDO profile quality. It can
be used to preserve local counts of functions which have 0 AFDO
profile.

I originally did not include it as it was not clear it will be useful and
it turns quality from 3bits to 4bits which means that we need to steal another
bit from the actual counters.

It is likely not a problem for profile_count since counting up to 2^60 still
takes a while.  However with profile_probability I run into a problem that
gimple FE testcases encodes probability with current representation and thus
changing profile_probability::n_bits would require updating all testcases.

Since probabilities never use GLOBAL0 qualities (they are always local)
addoing bits is not necessary, but it requires encoding quality in
the data type that required adding accessors.

While working on this I also noticed that GUESSED_GLOBAL0 and
GUESSED_GLOBAL0_ADJUSED are misordered. Qualities should be in
increasing order.  This is also fixed.

auto-profile will be updated later.

Bootstrapped/regtested x86_64-linux, comitted.

Honza

gcc/ChangeLog:

	* cgraph.cc (cgraph_node::make_profile_global0): Support
	GUESSED_GLOBAL0_AFDO
	* ipa-cp.cc (update_profiling_info): Use
	GUESSED_GLOBAL0_AFDO.
	* profile-count.cc (profile_probability::dump): Use
	quality ().
	(profile_probability::stream_in): Use m_adjusted_quality.
	(profile_probability::stream_out): Use m_adjusted_quality.
	(profile_count::combine_with_ipa_count): Use quality ().
	(profile_probability::sqrt): Likewise.
	* profile-count.h (enum profile_quality): Add
	GUESSED_GLOBAL0_AFDO; reoder GUESSED_GLOBAL0_ADJUSTED and
	GUESSED_GLOBAL0.
	(profile_probability): Add min_quality; replase m_quality
	by m_adjused_quality; add set_quality; update all users
	of quality.
	(profile_count): Set n_bits to 60; make m_quality 4 bits;
	update uses of quality.
	(profile_count::afdo_zero, profile_count::globa0afdo): New.