AArch64: propose -mmax-vectorization as an option to override vector costing
With the middle-end providing a way to make vectorization more profitable by
scaling vect-scalar-cost-multiplier this makes a more user friendly option
to make it easier to use.

I propose making it an actual -m option that we document and retain vs using
the parameter name.  In the future I would like to extend this option to modify
additional costing in the AArch64 backend itself.

This can be used together with --param aarch64-autovec-preference to get the
vectorizer to say, always vectorize with SVE.  I did consider making this an
additional enum to --param aarch64-autovec-preference but I also think this is
a useful thing to be able to set with pragmas and attributes, but am open to
suggestions.

Note that as a follow up I plan on extending -fdump-tree-vect to support -stats
which is then intended to be usable with this flag.

gcc/ChangeLog:

	* config/aarch64/aarch64.opt (max-vectorization): New.
	* config/aarch64/aarch64.cc (aarch64_override_options_internal): Save
	and restore option.
	Implement it through vect-scalar-cost-multiplier.
	(aarch64_attributes): Default to off.
	* common/config/aarch64/aarch64-common.cc (aarch64_handle_option):
	Initialize option.
	* doc/extend.texi (max-vectorization): Document attribute.
	* doc/invoke.texi (max-vectorization): Document flag.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/sve/cost_model_17.c: New test.
	* gcc.target/aarch64/sve/cost_model_18.c: New test.