rtl-ssa: Rewrite process_uses_of_deleted_def [PR120745]
process_uses_of_deleted_def seems to have been written on the assumption
that non-degenerate phis would be explicitly deleted by an insn_change,
and that the function therefore only needed to delete degenerate phis.
But that was inconsistent with the rest of the code, and wouldn't be
very convenient in any case.

This patch therefore rewrites process_uses_of_deleted_def to handle
general phis.

I'm not aware that this fixes any issues in current code, but it is
needed to enable the rtl-ssa dce work that Ondřej and Honza are
working on.

gcc/
	PR rtl-optimization/120745
	* rtl-ssa/changes.cc (process_uses_of_deleted_def): Rewrite to
	handle deletions of non-degenerate phis.