c++: Add fix note for how to declare main in a module
This patch adds a note to help users unfamiliar with modules terminology
understand how to declare main in a named module since P3618.

There doesn't appear to be an easy robust location available for "the
start of this declaration" that I could find to attach a fixit to, but
the explanation should suffice.

gcc/cp/ChangeLog:

	* decl.cc (grokfndecl): Add explanation of how to attach to
	global module.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>