Fortran: Ensure finalizers are created correctly [PR120637]
Finalize_component freeed an expression that it used to remember which
components in which context it had finalized already.  While it makes
sense to free the copy of the expression, if it is unused, it causes
issues, when comparing to a non existent expression. This is now
detected by returning true, when the expression has been used.

	PR fortran/120637

gcc/fortran/ChangeLog:

	* class.cc (finalize_component): Return true, when a finalizable
	component was detect and do not free it.

gcc/testsuite/ChangeLog:

	* gfortran.dg/asan/finalize_1.f90: New test.