arc: testsuite: Scan rlc instead of mov.hs
Due to the patch by Roger Sayle,
09881218137f4af9b7c894c2d350cf2ff8e0ee23, which introduces the use of
the `rlc rX,0` instruction in place of the `mov.hs`, the add overflow
test case needs to be updated.  The previous test case was validating
the `mov.hs` instruction, but now it must validate the `rlc`
instruction as the new behavior.

gcc/testsuite/ChangeLog:

	* gcc.target/arc/overflow-1.c: Replace mov.hs with rlc.

Signed-off-by: Luis Silva <luiss@synopsys.com>