ranger: Mark several member functions as final override
When GCC is built with clang, it emits warnings that several member
functions of various ranger classes override a virtual function of an
ancestor but are not marked with the override keyword.  After
inspecting the cases, I found that all these classes had other member
functions marked as final override, so I added the final keyword
everywhere too.

In some cases other such overrides were not explicitly marked as
virtual, which made formatting easier.  For that reason and also for
consistency, in such cases I removed the virtual keyword from the
functions I marked as final override too.

gcc/ChangeLog:

2025-06-24  Martin Jambor  <mjambor@suse.cz>

	* range-op-mixed.h (class operator_plus): Mark member function
	overflow_free_p as final override.
	(class operator_minus): Likewise.
	(class operator_mult): Likewise.
	* range-op-ptr.cc (class pointer_plus_operator): Mark member
	function lhs_op1_relation as final override.
	* range-op.cc (class operator_div::): Mark member functions
	op2_range and update_bitmask as final override.
	(class operator_logical_and): Mark member functions fold_range,
	op1_range and op2_range as final override.  Remove unnecessary
	virtual.
	(class operator_logical_or): Likewise.
	(class operator_logical_not): Mark member functions fold_range and
	op1_range as final override.  Remove unnecessary virtual.
	formatting easier.
	(class operator_absu): Mark member functions wi_fold as final
	override.